… packages/story-editor/src/components/panels/design/sizePosition/sizePosition.js