## Context
This code looked into the design panel performance.
<!-- What do we want to achieve with this PR? Why did we write this code? -->

## Summary
This PR contains the second of two separate failed approaches to increase perf in the design panel. Documenting both approaches here to help guide future approaches.

### Culling Unnecessary Re-renders From the Top Down (Original Approach)
This was the first approach I tried, and the code isn't documented in this PR. My goal here was to pass around and interact with the story data a little more wisely to avoid unnecessary re-renders. The main blocker I ran into here was the current architecture inhibiting this.

The main issues stem from 2 architectural choices that make the entire style pane deeply intertwined with selected element state:
1. Holding & staging element updates in local state before dispatching the updates to the top level story reducer. This is used primarily to leverage presubmit handlers, but might be subject to change with #6690 . I modeled out the update logic here in #11389
2. Having an abstract concept of a Design Panel. Since every input panel is abstracted on to be handled as a "design panel", we don't know how the design panel is going to consume the data. Because of this, we end up passing every design panel far more data than is essential for its function (ie `selectedElements` & `selectedElementAnimations`) even tho we rarely need all that info for the all inputs (ie color input only cares about color of current selection).

Altering the current architecture seemed out of scope for this ticket, and without doing that, I was unable to find any solutions that fixed performance issues by just improving how we pass around data.

### Culling Unnecessary Re-renders from the Bottom Up (Second Approach)
This is the approach outlined in the code changes in this PR. The basic idea is to reduce unnecessary re-renders internally in each panel by making each panel more or a wrapper component, inferring the data needed for the input in the wrapper, then making the panel's markup memoized and passing the inferred input data to the memoized input markup.

I tried out this approach on the alignment & sizePosition panels as a POC. The approach was successful in eliminating re-renders, but the eliminated re-renders didn't outweigh cost of the memoization. The memoization also added complexity to the code. Here are my findings:

#### Updating selection between copies of the same element (only position value changes)
**React Timing**
We saw an improvement here, but this doesn't provide the full picture. This isn't accounting for the cost of the memoization which in some cases actually made this slower, while in others it was faster, however basically averaged out to be about the same as before.

| main | this PR |
| ----- | ----- |
| ~11.5ms | ~6.5ms |
| <img width="691" alt="Screen Shot 2022-04-28 at 12 35 43 PM" src="https://user-images.githubusercontent.com/35983235/165813726-96b43a2f-a451-460e-92ca-967cee16ef8b.png"> | <img width="693" alt="Screen Shot 2022-04-28 at 12 39 36 PM" src="https://user-images.githubusercontent.com/35983235/165814315-4be736c7-7b4f-40d9-b535-fb32c011c174.png"> |

**Main Thread Timing**
Basically averages out to be about the same here and doesn't warrant the added technical complexity of making panels follow a wrapper & memoized controls model:

| main | this PR |
| ----- | ----- |
| ~44ms | ~44ms |
| <img width="692" alt="Screen Shot 2022-04-28 at 12 46 44 PM" src="https://user-images.githubusercontent.com/35983235/165815550-959464e0-2842-42d4-b94e-cc9ee377f7e8.png"> |  <img width="693" alt="Screen Shot 2022-04-28 at 12 43 58 PM" src="https://user-images.githubusercontent.com/35983235/165815578-22eeab49-789d-42d9-b8d0-69351daf77ea.png"> |

### Summary
I think if we really want to see more substantial performance improvements here, it would be in altering the design panels more substantially in #6690 , and there are no quicker wins here.

That being said, this interaction is already under our 100ms budget [outlined here](https://github.com/GoogleForCreators/web-stories-wp/blob/782fd71d2f74752a62ea69fe1173f530b8f85664/docs/architecture.md?plain=1#L22-L23), so it may not be worth investing more time in until we add new features/regressions that push us over our allotted 100ms of response time (main is currently ~44ms in react development build and ~22ms in react prod build on my macbook pro rn).

I think this ticket, #10072 ,  was more of a discovery ticket than anything if you read the description, so I think this is enough to close it out. Will close the PR too, but just wanted to have the code up here to reference in case anybody wanted to see the technical implementation of some of the tactics used.


## Relevant Technical Choices
Described above
<!-- Please describe your changes. -->

## To-do
NA
<!-- A list of things that need to be addressed in this PR or follow-up changes. -->

## User-facing changes
NA
<!--
Please describe your changes.
Include before/after screenshots or a short video.
-->

## Testing Instructions
NA
<!--
How can the changes in this PR be verified?
Please provide step-by-step instructions how to reproduce the issue, if applicable.
Write step-by-step test instructions aimed at non-tech-savvy users, even if the PR is not user-facing.
-->

<!-- ignore-task-list-start -->
- [ ] This is a non-user-facing change and requires no QA
<!-- ignore-task-list-end -->

This PR can be tested by following these steps:

1.


## Reviews

### Does this PR have a security-related impact?

<!-- Examples: new APIs, changes to KSES, etc.  -->

### Does this PR change what data or activity we track or use?

<!-- Examples: changes to telemetry, new third-party APIs -->

### Does this PR have a legal-related impact?

<!-- Examples: new images with unknown sources, new production dependencies with incompatible licenses -->

## Checklist

<!-- Check these after PR creation -->

- [x] This PR addresses an existing issue and I have linked this PR to it in ZenHub
- [ ] I have tested this code to the best of my abilities
- [ ] I have verified accessibility to the best of my abilities ([docs](https://github.com/googleforcreators/web-stories-wp/blob/main/docs/accessibility-testing.md))
- [ ] I have verified i18n and l10n (translation, right-to-left layout) to the best of my abilities
- [ ] This code is covered by automated tests (unit, integration, and/or e2e) to verify it works as intended ([docs](https://github.com/googleforcreators/web-stories-wp/tree/main/docs#testing))
- [ ] I have added documentation where necessary
- [x] I have added a matching `Type: XYZ` label to the PR

---

<!--
Please reference the issue(s) this PR addresses.
No URLs, just the issue numbers.
Use "Fixes #123" if it fixes an issue.

NOTE: One reference per line!

Example:

Fixes #123
Partially addresses #456
See #789
-->

Fixes #10072